requirements.txt
python-telegram-bot==20.0
apscheduler==3.8.1
pytz==2025.2
